refactor(bash): shorten scripts

+++ res/init/init.sh
@@ -5,7 +5,7 @@ cd $(dirname $0)
 . ../util/os.sh
 
 # install Homebrew
-if type brew >/dev/null 2>&1; then
+if type brew &>/dev/null; then
   echo "${BLUE}Info:${RESET} Homebrew is already installed."
 else
   declare os_name=$(get_os_name)
@@ -30,22 +30,8 @@ fi
 brew doctor
 
 # install Homebrew formulae
-if [ ! -f "$(brew --prefix)/etc/bash_completion" ]; then
-  brew install bash-completion
-fi
-
-if !(type dircolors >/dev/null 2>&1 || type gdircolors >/dev/null 2>&1); then
-  brew install coreutils
-fi
-
-if !(type git >/dev/null 2>&1); then
-  brew install git
-fi
-
-if !(type ghq > /dev/null 2>&1); then
-  brew install ghq
-fi
-
-if !(type fzf > /dev/null 2>&1); then
-  brew install fzf
-fi
+[ ! -f "$(brew --prefix)/etc/bash_completion" ] && brew install bash-completion
+!(type dircolors &>/dev/null || type gdircolors &>/dev/null) && brew install coreutils
+!(type git &>/dev/null) && brew install git
+!(type ghq &>/dev/null) && brew install ghq
+!(type fzf &>/dev/null) && brew install fzf
